Appendix: Change linux git clone location
The training uses the local copy of that repo, so it may be confusing to point
reader the how to get the copy. It may look like it's an important step.

Althought it is not an important step to the traning. It is very important
information, so it's included in the Appendix.

Signed-off-by: Daiane Angolini <daiane.angolini@nxp.com>